fixed usergen @vrde
