+++ lib/adopters.js
@@ -6,6 +6,29 @@ const nconf = require('nconf');
 const mongo = require('./mongo');
 const utils = require('../lib/utils');
 
+const requiredHeaders =  {
+    'content-length': 'length',
+    'x-fbtrex-build': 'build',
+    'x-fbtrex-version': 'version',
+    'x-fbtrex-userid': 'supporterId',
+    'x-fbtrex-publickey': 'publickey',
+    'x-fbtrex-signature': 'signature'
+};
+
+function processHeaders(received) {
+    var ret = {};
+    var errs = _.map(requiredHeaders, function(destkey, headerName) {
+        var r = _.get(received, headerName);
+        if(_.isUndefined(r))
+            return headerName;
+
+        _.set(ret, destkey, r);
+        return null;
+    });
+    errs = _.compact(errs);
+    return _.size(errs) ? { errors: errs } : ret;
+};
+
 function create(headers) {
     // when this function is called we are sure the 
     // combo 'userId'+'publicKey' do not exists
@@ -26,7 +49,7 @@ function create(headers) {
         random: _.random(0, supporter.userId),
         when: moment().toISOString()
     });
-    debug("Creating %s (%d)", supporter.pseudo, supporter.userId);
+    debug("Creating %s %s (%d)", supporter.pseudo, supporter.publicKey, supporter.userId);
     return mongo
         .writeOne(nconf.get('schema').supporters, supporter)
         .return(supporter);
@@ -56,4 +79,5 @@ function validateToken(userToken) {
 module.exports = {
     create: create,
     validateToken: validateToken,
+    processHeaders,
 };
